requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
scikit-learn==1.4.1.post1
sentence-transformers==2.5.1
faiss-cpu==1.8.0
numpy==1.26.4
pydantic==2.6.4
